+++ internal/platform/providers/redataapi/responses.go
@@ -4,7 +4,7 @@ type fetchPVPCPricesResponse struct {
 	Included []struct {
 		Attributes struct {
 			Values []struct {
-				Value    float32 `json:"value"`
+				Value    float64 `json:"value"`
 				Datetime string  `json:"datetime"`
 			} `json:"values"`
 		} `json:"attributes"`
